Update scripts and README

+++ build.sh
@@ -1,6 +1,14 @@
 #!/bin/sh -e
 
+# install dependencies, run unit tests, and bundle all the files into a deployable zip file.
+
+if [ ! -f ./config.json ]; then
+    echo "Unable to build Lambda.zip! \"config.json\" is required!"
+    exit 1
+fi
+
 npm install
+
 
 rm -rf ./dist
 
@@ -8,4 +16,4 @@ mkdir -p dist
 
 npm test
 
-zip -r -q dist/lambda.zip ./ -x ./*\.sh ./.git/**\* ./dist 
+zip -r -q dist/lambda.zip ./ -x ./*\.sh ./.git/**\* ./dist ./spec ./scripts ./.idea
